FIRE-34893 - ensure model preview is cleanly initialised
Crash appears to be due to an attempt to access the modelloader thread after it has been destroyed. The handle to the preview is passed in (to a number of places) as part of a callback if the crashes reappear then we might need to have a weak_ptr on the preview itself to test validity.